Move enclosinFunction to util
